requirements.txt
streamlit
pandas
numpy
matplotlib
requests
seaborn
plotly
cloudinary
python-dotenv
streamlit_extras

gspread
oauth2client
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
